moved package jgfx.javagradlefx -> testing
