+++ src/test/java/testing/FileParserTest.java
@@ -1,4 +1,4 @@
-package jgfx.javagradlefx;
+package testing;
 
 import app.controller.FileParser;
 import app.controller.Settings;
