+++ src/test/java/testing/RayTest.java
@@ -1,4 +1,4 @@
-package jgfx.javagradlefx;
+package testing;
 
 import app.controller.graphicsEngine.Ray;
 import app.controller.linAlg.Vector;
